Final ellipse Result
Unexpectedly the program ran with no errors and the ellipse appeared as if the arguments were still +80. Switched back to +80 to see if there was any difference in the ellipse size and there appeared to be none. Processing possibly views negative numbers equally to positive numbers when entered as a method's arguments.